% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awra2020data.R
\docType{package}
\name{awra2020spatial}
\alias{awra2020spatial}
\alias{awra2020spatial-package}
\title{Sample datasets to use for AWRA 2020 R spatial workshop}
\description{
This package contains sample data to use for
demonstration purposes for the AWRA 2020 R spatial workshop.
}
