% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awra2020data.R
\docType{data}
\name{bike_paths}
\alias{bike_paths}
\title{Austin bike paths - a dataset of bicycle transportation
facilities developed by the City of Austin and published
as an ArcGIS REST service.}
\format{An sf data frame of 12203 observations and 20 variables}
\source{
\url{https://services.arcgis.com/0L95CJ0VTaxqcmED/arcgis/rest/services/TRANSPORTATION_bicycle_facilities/FeatureServer/0}
}
\usage{
bike_paths
}
\description{
Austin bike paths - a dataset of bicycle transportation
facilities developed by the City of Austin and published
as an ArcGIS REST service.
}
\keyword{datasets}
