% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awra2020data.R
\docType{data}
\name{parks}
\alias{parks}
\title{Austin parks - a dataset of park boundaries developed by
the City of Austin and publishedas an ArcGIS REST service.}
\format{An sf data frame of 304 observations and 56 variables}
\source{
\url{https://services.arcgis.com/0L95CJ0VTaxqcmED/arcgis/rest/services/BOUNDARIES_city_of_austin_parks/FeatureServer/0}
}
\usage{
parks
}
\description{
Austin parks - a dataset of park boundaries developed by
the City of Austin and publishedas an ArcGIS REST service.
}
\keyword{datasets}
